Implements grouped removals
MutationRecords will contain groupings of removed and added nodes. It
looks something like:

[
	{removedNodes: [1]},
	{removedNodes: [2]},
	{removedNodes: [3]},
	{addedNodes: [14]},
    {addedNodes: [15]},
	{removedNodes: [5]},
	{removedNodes: [6]},
	{addedNodes: [13]}
]

This commit ensures that we process removed nodes in reverse order, so
that the indices stay correct. In the above example we will:

1. Remove node 3
2. Remove node 2
3. Remove node 1
4. Add node 14
5. Add node 15
6. Remove node 6
7. Remove node 5
8. Add node 13

It's possible that we should be going in another order (6,5,3,2,1), and
actually doing *all* removals for the entire records set in reverse
order, but I haven't ran into that sort of scenario yet.