Test using ir-clone
This changes are tests for consecutive TextNodes to use ir-clone. This
is a tool that can serialize a document and insert comment nodes between
textnodes, to preserve them when IR syncing occurs.

The next part of this change will be to remove collapseTextNodes option
as it will no longer be needed. That will occur in another commit so
that it can be undone if necessary.